Namespace refactoring in integration tests
