+++ Neo4j.Driver/Neo4j.Driver.Tests.Integration/Direct/AuthenticationIT.cs
@@ -20,11 +20,10 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using FluentAssertions;
 using Neo4j.Driver.Internal;
-using Neo4j.Driver;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace Neo4j.Driver.IntegrationTests
+namespace Neo4j.Driver.IntegrationTests.Direct
 {
     public class AuthenticationIT : DirectDriverTestBase
     {
