+++ Neo4j.Driver/Neo4j.Driver.Tests.Integration/Direct/BoltTypesIT.cs
@@ -22,7 +22,7 @@ using FluentAssertions;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace Neo4j.Driver.IntegrationTests
+namespace Neo4j.Driver.IntegrationTests.Direct
 {
     public class BoltTypesIT : DirectDriverTestBase
     {
