+++ Neo4j.Driver/Neo4j.Driver.Tests.Integration/Direct/TransactionIT.cs
@@ -17,15 +17,12 @@
 
 using System;
 using System.Diagnostics;
-using System.IO;
-using System.Linq;
 using System.Threading.Tasks;
 using FluentAssertions;
-using Neo4j.Driver;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace Neo4j.Driver.IntegrationTests
+namespace Neo4j.Driver.IntegrationTests.Direct
 {
     public class TransactionIT : DirectDriverTestBase
     {
