+++ Neo4j.Driver/Neo4j.Driver.Tests.Integration/Routing/RoutingDriverTestBase.cs
@@ -18,12 +18,10 @@
 using System;
 using Neo4j.Driver.IntegrationTests.Internals;
 using Neo4j.Driver.IntegrationTests.Shared;
-using Neo4j.Driver.Internal;
-using Neo4j.Driver;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace Neo4j.Driver.IntegrationTests
+namespace Neo4j.Driver.IntegrationTests.Routing
 {
     [Collection(CCIntegrationCollection.CollectionName)]
     public abstract class RoutingDriverTestBase : IDisposable
