+++ Neo4j.Driver/Neo4j.Driver.Tests.Integration/Types/TemporalTypesIT.cs
@@ -22,6 +22,7 @@ using System.Threading.Tasks;
 using FluentAssertions;
 using Neo4j.Driver.Internal;
 using Neo4j.Driver;
+using Neo4j.Driver.IntegrationTests.Direct;
 using Xunit;
 using Xunit.Abstractions;
 using static Neo4j.Driver.IntegrationTests.VersionComparison;
